+++ API/Models/Models.go
@@ -70,20 +70,11 @@ type PreliminarySchedule struct {
 	Sessions []PreliminarySession `json:"preliminary_sessions"`
 }
 
-type Login struct {
-	Identifier int    `json:"identifier"`
-	Username   string `json:"username"`
-	Type       int    `json:"type"`
-	Password   string `json:"-"`
-	Token      string `json:"token"`
-}
-
 type Client struct {
-	MembershipNumber 	int 	`json:"membershipNumber"`
-	Name           		string 	`json:"name"`
-	Email          		string 	`json:"email"`
-	Phone          		string 	`json:"phone"`
-	Balance        		string 	`json:"balance"`
-	Identification 		string 	`json:"identification"`
+	MembershipNumber int    `json:"membershipNumber"`
+	Name             string `json:"name"`
+	Email            string `json:"email"`
+	Phone            string `json:"phone"`
+	Balance          string `json:"balance"`
+	Identification   string `json:"identification"`
 }
-
