+++ API/Models/User.go
@@ -15,10 +15,9 @@ type User struct {
 }
 
 type Login struct {
-	ID             int    `json:"id"`
-	Username       string `json:"username"`
-	Type           int    `json:"type"`
-	Identification string `json:"identification"`
-	Password       string `json:"-"`
-	Token          string `json:"token"`
+	Identifier int    `json:"identifier"`
+	Username   string `json:"username"`
+	Type       int    `json:"type"`
+	Password   string `json:"-"`
+	Token      string `json:"token"`
 }
