+++ API/WebServer/Controllers/clientController.go
@@ -2,7 +2,7 @@ package Controllers
 
 import (
 	"API/Database/Requests"
-	"API/WebServer/Token"
+	"strconv"
 
 	"github.com/gofiber/fiber/v2"
 )
@@ -15,9 +15,15 @@ func GetClientInfo(context *fiber.Ctx) error {
 		return nil
 	}
 
-	registeredUser := Token.GetUsernameFromToken(token)
+	var data map[string]string
+	if err := context.BodyParser(&data); err != nil {
 
-	user := Requests.GetClientProfileInfo(registeredUser)
+		return err
+	}
+
+	membershipNumber, _ := strconv.Atoi(data["membershipNumber"])
+
+	user := Requests.GetClientProfileInfo(membershipNumber)
 
 	return giveJSONResponse(context, user, fiber.StatusOK)
 }
