+++ API/WebServer/routes.go
@@ -8,7 +8,6 @@ import (
 
 func Setup(app *fiber.App) {
 	app.Get("/", Controllers.Start)
-	app.Get("/test", Controllers.SqlTests)
 
 	general := app.Group("/general")
 	client := app.Group("/client")
@@ -34,5 +33,38 @@ func Setup(app *fiber.App) {
 	admin.Post("/removeInstructor", Controllers.DeleteInstructor)
 
 	admin.Post("/cancelSession", Controllers.CancelSession)
+
+}
+
+func Setup2(app *fiber.App) {
+	app.Get("/", Controllers.Start)
+
+	client := app.Group("/client")
+	client.Post("/reservedSessions", Controllers.GetReservedSessions)
+
+	// ! Esta ruta retorna usuario cliente
+	client.Get("/userInfo", Controllers.GetClientInfo)
+
+	user := app.Group("/user")
+	user.Post("/login", Controllers.Login)
+
+	services := app.Group("/services")
+	services.Get("/list", Controllers.GetServices)
+
+	sessions := app.Group("/sessions")
+	sessions.Get("/activeSchedule", Controllers.GetActiveSchedule)
+	sessions.Post("/bookSession", Controllers.BookSession)
+	sessions.Post("/cancelBooking", Controllers.CancelBooking)
+	sessions.Post("/cancelSession", Controllers.CancelSession)
+
+	instructor := app.Group("/instructor")
+	instructor.Post("/list", Controllers.GetInstructors)
+	instructor.Post("/remove", Controllers.DeleteInstructor)
+
+	preliminarySchedule := app.Group("/preliminarySchedule")
+	preliminarySchedule.Post("/preliminarySchedule", Controllers.GetPreliminarySchedule)
+	preliminarySchedule.Post("/insertPreliminarySession", Controllers.InsertPreliminarySession)
+	preliminarySchedule.Post("/deletePreliminarySession", Controllers.DeletePreliminarySession)
+	preliminarySchedule.Post("/confirmPreliminarySchedule", Controllers.ConfirmPreliminarySchedule)
 
 }
